[#25439] DocDB: Fix potential segv issue while fetching txn status of sealed transaction
Summary:
While fetching the transaction status at the coordinator, if the transaction exists in the list of managed transactions, we access the iterator at a couple of places assuming we still hold the mutex. But looks like the lock can be released and re-acquired when dealing with sealed transactions, so accessing the earlier iterator after this release and re-acquisition of the lock isn't safe.

This diff fixes the issue by accessing the iterator before any potential release and re-acquire of the lock.
Jira: DB-14676

Test Plan: Jenkins

Reviewers: sergei

Reviewed By: sergei

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40889