[PLAT-16613][PLAT-16614] Restore with ON_ERROR_STOP and backup with roles
Summary:
To enable the restore with ON_ERROR_STOP, the database is going to
implement some protection around applying grants to roles that do not
exist. See https://github.com/yugabyte/yugabyte-db/issues/25877.

This adds a new flag to the backup config metadata that will let yba
know if it safe to pass on_error_stop based on if the database has the
above feature implemented.

In addition, Integrated with new YBC flags to allow backup and restore
of roles, with the api keys defaulted to off

Test Plan:
    wrote new UTs
    created universe, with some new tables/data:
```
    yugabyte=# \c postgres
    postgres=# create user shubin;
    CREATE ROLE
    postgres=# create role daniel;
    CREATE ROLE
    postgres=# grant all privileges on database postgres to shubin;
    GRANT
    postgres=# create table test1 (id int, name text);
    CREATE TABLE
    postgres=# insert into test1 values (0, 'test1');
    INSERT 0 1
    postgres=# grant all privileges on table test1 to shubin;
    GRANT
    postgres=# alter table test1 owner to daniel;
    ALTER TABLE
    postgres=#
```

    1. Backup with no extra args passed in. role checks existed
    2. restore with ignore errors = false: Passed, no alters/grants run
    3. backup without dump role checks: role checks did not exist
    4. Restore with use role. Auto added ignore errors = true
    5. restore with use roles and ignore_errors = false. ignore errors
       got overwritten to true.

Reviewers: vkumar, sanketh, loginov

Reviewed By: vkumar

Subscribers: svc_phabricator, steve.varnau, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41006