[BACKPORT 2024.1][#21775] docdb: Avoid segfault on catalog reload / collector call races.
Summary:
The existing `GetClusterConfig()` function can segfault if it is called while the catalog manager is reloading the system catalog because the `cluster_config_` field is reset during catalog loading. However it does already return a `Status` object, there is just no codepath that returns a not-OK status. This diff refactors the function to return a `Result` instead, and returns a not-OK status if the `cluster_config_` field is null. Since all clients are already dealing with the returned `Status` object in the old signature, this was a minimal change that makes calls safe, regardless of whether the caller holds the leader lease or not.
Jira: DB-10650

Original commit: 07da365ebde57435f113a98f513a8acf72da4e4d / D34609

Test Plan:
Existing tests to be sure the refactor is safe.

I didn't add a test that the race is fixed as the race is difficult to expose deterministically.

Reviewers: asrivastava, mhaddad, xCluster, hsunder

Reviewed By: asrivastava

Subscribers: slingam, esheng, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40985