[#25480] xCluster: Keep running xcluster task even if leader temporarily loses the lease
Summary:
On an xCluster target universe the yb-master leader computes the xCluster safe time every second. This task stops if the leadership is lost, and restarts again when a new leader is elected and sys catalog is reloaded. In some rare cases the master can lose the raft lease, but still retain its leadership. In this case we should keep running the xCluster task until we are certain that the leadership is lost.

`SCOPED_LEADER_SHARED_LOCK` checks if we are the leader and lease is valid. This change uses `CatalogManager::CheckIsLeaderAndReady` to check if the leadership is still valid.

Removing unnecessary `leader_term_` member.
Jira: DB-14730

Test Plan:
Jenkins

./yb_build.sh asan --cxx-test integration-tests_xcluster_ysql-test --gtest_filter XClusterYSqlTestConsistentTransactionsTest.AddServerIntraTransaction -n 40  --tp 1

Reviewers: sergei, jhe, slingam, xCluster

Reviewed By: sergei, jhe

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40988