[PLAT-16412] Auto node agent enabler shuts down executor service if the installation does not complete in 1min
Summary:
The method processNodes is run async in an executor service and is not supposed to return before the installation finishes unless there is an error/exception. Otherwise, the executor service is shutdown while the
installation is happening. This is reproducible easily when the installation takes long time more than a minute (about 5 mins on my local setup due to large node-agent packaging with pex). With this fix, it is not seen.

Test Plan:
1. Create a universe without node agent (client is disabled).
2. Enable it to let the background installer kick in.
3. Installation took very long due to large size of node-agent but it succeeded. Previously it used to fail on every attempt.

-rw-r--r--   1 nkhogen  staff   715M Jan  2 11:53 node_agent-2.25.1.0-PRE_RELEASE-linux-amd64.tar.gz

Reviewers: cwang, amalyshev

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41011