[#24294][YSQL] Fixing the test TestConcurrentOperations
Summary:
In `yb_oid_entry.c`, the fields database_count and user_count are accessed by multiple threads during backend connect, or while creating a new pool. The mutation of the above fields need to be serialized when the above methods are called. The change here is to utilize read-write locks to manage the access of the above fields in a thread-safe manner and avoiding potential race conditions.

The fix has been done in relation to the TestConcurrentOperations test failure in TestDropAndRenameRole and TestDropAndRenameDb

Test Plan:
Jenkins: enable connection manager, all tests

This change would fix the tests `TestDropAndRenameRole#TestConcurrentOperations` and `TestDropAndRenameDb#TestConcurrentOperations`

Reviewers: mkumar, rbarigidad, skumar, devansh.saxena, stiwary

Reviewed By: devansh.saxena

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D41022