[#24711] DocDB: Introduce session level docdb transaction for supporting session advisory locks
Summary:
To support session advisory locks, we need to somehow tag the locks/intents to an identifier which shares the same lifetime of the pg session. Additionally, we would need wait-on-conflict and deadlock detection with session level advisory lock conflicts as well. Hence, for each pg session requesting a session level advisory lock, we would need to maintain a docdb transaction that corresponds to the pg session.

This diff introduces support for creating a session level docdb transaction (if one doesn't already exist) on request for a session level advisory lock. This transaction is aborted on expiration of the session.

Note:
1. It could totally happen that this session level transaction gets aborted due to other reasons (missed heartbeats etc). If we want to honor the assumption that the session advisory locks are valid for the lifetime of the session, then we need to check the status of this transaction at every perform rpc.
2. Work on wait-on-conflict for session advisory lock conflicts and deadlock detection to follow in subsequent revisions.
Jira: DB-13787

Test Plan:
Jenkins

./yb_build.sh --cxx-test='TEST_F(PgAdvisoryLockTest, SessionAdvisoryLockAndUnlock) {'
./yb_build.sh --cxx-test='TEST_F(PgAdvisoryLockTest, CleanupSessionAdvisoryLock) {'

Reviewers: esheng, yyan, rthallam

Reviewed By: esheng, yyan

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D40932