[#25498] YSQL: Enable check_for_tables_with_oids
Summary:
In vanilla PG11, creating a table with OIDs (`CREATE TABLE t (i INTEGER) WITH OIDS;`) succeeds.  In YB PG11, this syntax was never supported, so the statement errored with `ERROR:  OIDs are not supported for user tables.`. (`CREATE TABLE ... WITH` was enabled as part of D6755 / 5df3abd, and `WITH OIDS` was disabled in that change).

Since there is no way to have a table created `WITH OIDS` in YB PG11, it's not a case we need to worry about. We could either enable the check (and it will never catch anything, but the code will be more consistent with Postgres), or disable it (saving a little bit of time on the pre-flight checks, but resulting in slightly more divergence from Postgres.)

Because it should be harmless to leave it in, and consistency with Postgres might be helpful for the next merge, I chose to re-enable the check.
Jira: DB-14757

Test Plan: Jenkins

Reviewers: fizaa, hsunder

Reviewed By: fizaa

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D41031