[#25379] docdb: Introduce per-transaction LockTracker state
Summary:
The current LockTracker implementation for ObjectLockManager involves a multi-index
container uniquely keyed on (versioned_txn, subtxn_id, object_id) and additionally indexed on
various subsets of the unique key. This means that we have to acquire the global
mutex an additional time on lock acquire/release/wait/wakeup.

This diff refactors this map to instead be keyed on versioned_txn and point to a per-transaction
struct (TrackedTransactionLockEntry) which contains a further
subtxn_id => object_id => TrackedLockEntry mapping. This allows us to pass around this new
TrackedTransactionLockEntry and avoid the need for a repeated global mutex lock for LockTracker
functionality.

TrackedTransactionLockEntry is currently only briefly passed around through the lock acquire path
to avoid repeated global mutex locks; storing it externally to help avoid the need for the initial
global mutex lock will be addressed in a later diff.

This is also needed for the planned move of ObjectLockManager into shared memory, to avoid
partially updated indexes in the multi-index in event of PG crashing while inserting/deleting from
the container.
Jira: DB-14609

Test Plan: Jenkins.

Reviewers: bkolagani

Reviewed By: bkolagani

Subscribers: sergei, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40814