[PLAT-16419] Revert YBA UI search bar token match to substring search instead of regex matching
Summary:
Context:
The search bar tokens are read as regex patterns, but this is not communicated clearly to the user. This
can cause issues like a crash if the user enters special characters like `(` without enclosing `)` or unexpected
matches if users input charcters sequneces like `.*` unaware that it will be read as regex.

Change:
We are reverting back to supporting only substring matches as this covers the vast majority of use cases.

Test Plan:
Navigate to the slow query page and create search tokens using various regex patterns which make use of
special characters. verify that these tokens are treated as simple strings and the results are from substirng
match.
{F323306}
{F323308}
{F323307}

Reviewers: kkannan, rmadhavan

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41093