[PLAT-16433] disable the yba minimum version check when adding a release
Summary:
This check was originally added when YBA was possibly going to get a different
version scheme. That has not happened yet, so we do not currently need the check.

This will overwrite yb.skip_version_checks or yb.allow_db_version_more_than_yba_version

Test Plan:
tested adding release with the check enabled. it failed.
disabled the check, release was added

Reviewers: muthu

Reviewed By: muthu

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41100