[PLAT-16435] Update search placeholder text for restore backup modal
Summary:
The search placeholder text for searching keyspaces is `Search table name`.
The component here takes the user input and filters the list of keyspaces shown.
As a result, the search placeholder text should be renamed to 'Search keyspace name' instead
to reflect this logic.

Test Plan:
Verify the text change is applied as expected.
{F323325}

Reviewers: rmadhavan, kkannan

Reviewed By: rmadhavan, kkannan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41101