[#24108] CDC: Add column yb_lsn_type to pg_replication_slots view
Summary:
This diff adds a column for `yb_lsn_type` to the `pg_replication_slots` view.

Additionally, a migration script has also been added to facilitate upgrades.

**Upgrade / rollback safety:**

This diff adds a field `yb_lsn_type` to the message `PgReplicationSlotInfoPB`.

This revision does not alter persisted data. New fields are added to the tserver response (to pggate) proto. When the value is absent, master (which is upgraded first) is expected to fill in the appropriate default value. This is upgrade and rollback safe. No new flags are added to guard the feature.
Jira: DB-12997

Test Plan: Run existing tests.

Reviewers: skumar, stiwary, sumukh.phalgaonkar, aagrawal, utkarsh.munjal, xCluster, hsunder

Reviewed By: aagrawal, utkarsh.munjal

Subscribers: yql, ybase, ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D38846