[PLAT-16420]: Fix List CDC replication slot API
Summary:
Fixes list CDC replication slot api.
There could be some CDC replication slot in deleting state which has their namespace in deleting state
and if a namespace is in deleting state, it is not listed in namespace list api which shows no record.
Made the changes to only populate the namespace in slot details only if it exists.

Test Plan: added unit test

Reviewers: asharma

Reviewed By: asharma

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41130