[#25042] YSQL: Fix yb_lock_status test for ASAN/TSAN
Summary:
`yb_lock_status` in `testPgRegressProc` has been failing in ASAN and TSAN consistently.
The cause of failure is a change in output between PG 11 and PG 15 for queries of the form: `SELECT <constant> FROM <relation>;`
In PG 11, this produced an output in the following format:
```
yugabyte=# SELECT true FROM test;
 bool
------
(0 rows)
```
while in PG 15, the output was as follows:
```
yugabyte=# SELECT true FROM test;
 ?column?
----------
(0 rows)
```
In other words, in PG 15, the name of the column is no longer set to be the data type of the constant, if the data type is not explicitly specified.

D31424 fixed the output in the PG 15 branch for the output files: `yb_lock_status.out` and `yb_lock_status_1.out`
Later, D32492 introduced a new output file with the PG 11 style output: `yb_lock_status_2.out`

The files correspond to the following build types/configurations:
 - yb_lock_status_2.out: ASAN, TSAN
 - yb_lock_status_1.out: All other build types
 - yb_lock_status.out: All build types when READ COMMITTED isolation level is not enabled.

This revision modifies the queries to explicitly specify the data type for queries of the form `SELECT <constant> FROM <relation>;`
This will also prevent the test from failing again if the name of the column in the output changes again in the future.

Note that `testPgRegressProc` will continue to fail on Mac due to a different regress test (yb_get_current_transaction_priority).
Jira: DB-14176

Test Plan:
Run the following test:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressProc#testPgRegressProc'
```

Reviewers: patnaik.balivada

Reviewed By: patnaik.balivada

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D41136