[PLAT-16444] Update YBA documentation links for xCluster schema change mode
Summary:
We have a new documentation page which covers the differences between the schema change modes and how to upgrade
from manual to semi-automatic.
This diff updates the documetation links on the upgrade info modal to surface this information.

This diff also contains minor text updates.

Test Plan:
Create an xCluster replication config without enabling semi-automatic mode.
Verify the documentation links on the upgrade info modal is as expected and the text changes
are present.
{F323834}

Reviewers: hzare, ghodge

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41137