[PLAT-16152] Correctly delete PITR configs on A -> B and A -> C xcluster config setups with same db
Summary:
Allows PITR config to be correctly deleted for DR/xcluster configurations with A -> B and A -> C setups using the same databases. The databases on A will use the same PITR configs for both replications. Hence, make sure that if A -> B or A -> C is deleted, the pitr configs on A are retained and not deleted.

Note that there are still issues with A -> B and A -> C if the pitr params are different which cannot be solved with the current implementation.

Test Plan:
UTs

Manually tested following flow:
Create A -> B and A -> C with the same database and pitr params. Check that there is one pitr config for the database on A. Delete A -> C replication. Validate that the pitr conifg

Reviewers: hzare, jmak

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41172