[#25550] CDC: Fix bug causing stuck lag metric after table becomes not of interest
Summary:
We are using an internal structure `tablet_checkpoints` to keep track of the checkpoints. This structure contains a list of all the tablets which have been polled so far.

Additionally, we find the expired or tablets not of interest and store it in a list `expired_entries`. Now to find the metric objects that we need to clean, we have the following logic:

```
for (auto checkpoint : tablet_checkpoints) {
  if (checkpoint.tablet is present in expired_entries) {
    ResetMetrics();
  }
}
```

The bug here is that for unconsumed slots, we will never be getting an entry in `tablet_checkpoints` and in turn we will never reset the metrics. This diff fixes the issue by adding a dummy entry to the `tablet_checkpoints` if an entry for an expired or not of interest tablet is not present.
Jira: DB-14742

Test Plan:
Added a new test to validate the fix:

```
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestLagMetricForUnconsumedSlot
```

Reviewers: sumukh.phalgaonkar, siddharth.shah, skumar

Reviewed By: sumukh.phalgaonkar

Subscribers: ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D41116