[#25501] YSQL: Do not run check_new_cluster when the new cluster is not reachable
Summary:
`check_new_cluster` connects to the new cluster to grab information about the databases and user tables present on the new cluster. The C++ tests did not run a preflight check, so they missed this case. Adding the preflight checks reproduced the same issue that the shell test ran into.

When `pg_upgrade --check` is run before any other upgrade step, all we have is a PG11 cluster. So its not possible to connect to the new cluster to run checks on it, but pg_upgrade expects to be able to do so.

However, when pg_upgrade is run during the upgrade, the new cluster is reachable. So rather than disabling the check entirely, the check is disabled only if the new cluster is not reachable.
Jira: DB-14760

Test Plan:
Jenkins
```
pg15_tests/test_upgrade.sh
./yb_build.sh release --cxx-test integration-tests_pg15_upgrade-test --gtest_filter Pg15UpgradeTest.SimpleTableUpgrade
```

Reviewers: hsunder, fizaa

Reviewed By: hsunder

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D41042