[#25551] YSQL: Use yugabyte_upgrade user when upgrading with non local tserver
Summary:
When there is no local yb-tserver running on the yb-master node, and the universe has auth enabled, pg_upgrade and pg_restore have to use a custom user `yugabyte_user` with password auth. The password is expected to be stored on the `.pgpass` file in the yb-master nodes.
The restore is always done as the `yugabyte` user, since that is the only user that exists on the new pg15 catalog before the restore.
The backup for a binary upgrade typically skips the current user running the command, but in our case the `yugabyte_user` runs the backup and `yugabyte` runs the restore, we create all users with a check for existence.

This change also allows the root user to run pg_upgrade, which is required for kubernetes deployments.
Switched the pg_upgrade Subprocess to send out and err to standard location instead of capturing it in a string since the capture does not work when the result is too big, and causes the process to hang.

Fixes #25551
Jira: DB-14803

Test Plan:
Pg15UpgradeTestWithAuth.NoTserverOnMasterNode

And manually tested with `.pgpass` file:

```
# Start old cluster with auth enabled and expression pushdown disabled.
pushd /tmp/yugabyte-2024.2.1.0
./bin/yb-ctl wipe_restart --rf 3 --tserver_flags='ysql_enable_auth=true,ysql_pg_conf_csv=yb_enable_expression_pushdown=false' --master_flags='ysql_enable_auth=true'
# Create the upgrade user
./bin/ysqlsh -h /tmp/.yb.127.0.0.1:5433 -c "CREATE USER yugabyte_upgrade WITH SUPERUSER PASSWORD 'yugabyte'"
echo "*:*:*:yugabyte_upgrade:yugabyte" >> ~/.pgpass

popd

# Run pg_upgrade --check as root user
sudo ./build/latest/postgres/bin/pg_upgrade --check --old-host 127.0.0.1 --old-port 5433 --username yugabyte_upgrade --old-datadir /Users/hsunder/yugabyte-data/node-1/disk-1/pg_data_11

# Upgrade yb-masters
./bin/yb-ctl restart_node 1 --master --master_flags='ysql_enable_auth=true'
./bin/yb-ctl restart_node 2 --master --master_flags='ysql_enable_auth=true'
./bin/yb-ctl restart_node 3 --master --master_flags='ysql_enable_auth=true'
# Shutdown tserver on the master leader
./build/latest/bin/yb-admin --master_addresses 127.0.0.1:7100,127.0.0.2:7100,127.0.0.3:7100 list_all_masters
./bin/yb-ctl stop_node 1
sleep 65 # Wait for node to be marked as dead
./build/latest/bin/yb-admin --master_addresses 127.0.0.1:7100,127.0.0.2:7100,127.0.0.3:7100 list_all_tablet_servers | grep DEAD

# Run the upgrade
./build/latest/bin/yb-admin --master_addresses 127.0.0.1:7100,127.0.0.2:7100,127.0.0.3:7100 ysql_major_version_catalog_upgrade

grep yugabyte_upgrade ~/yugabyte-data/node-1/disk-1/yb-data/master/logs/yb-master.INFO
# I0110 11:13:44.465322 1903030272 ysql_initdb_major_upgrade_handler.cc:361] Running ysql major upgrade on a authentication enabled universe which does not have a yb-tserver on the same node as the yb-master. Upgrade will be performed using yb-tserver hosted on 127.0.0.2:5433by user yugabyte_upgrade
# I0110 11:13:44.465591 1903030272 pg_wrapper.cc:930] Launching pg_upgrade: [/Users/hsunder/code/yugabyte-db/build/latest/postgres/bin/pg_upgrade, --new-datadir, /Users/hsunder/yugabyte-data/node-1/disk-1/pg_data, --username, yugabyte_upgrade, --new-socketdir, /tmp/.yb.127.0.0.1:5434, --new-port, 5434, --old-port, 5433, --old-host, 127.0.0.2]
# Success!

```

Reviewers: telgersma, smishra

Reviewed By: telgersma

Subscribers: anijhawan, vbansal, ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D41171