[#25554] YSQL: Rid off multiple code blocks for pg_types map building
Summary:
There are multiple code blocks which build map of YugaByte and PostgreSQL datatypes.

`PgTypeInfo` struct does this and also `PgApiImpl` does the same thing. It is reasonable to re-use `PgTypeInfo` instead.

**Note:**
In context of this diff some cleanup changes are made. For example the `PgApiContext` is removed.
Jira: DB-14809

Test Plan: Jenkins

Reviewers: pjain, tnayak, telgersma, myang, jason

Reviewed By: myang

Subscribers: ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D40931