[#24077] xCluster: add new (currently disabled) test for replicating enum DDLs
Summary:
Here I add a test, XClusterPgRegressDDLReplicationTest.PgRegressCreateDropExtensions, to ensure we can automatically replicate enum DDLs.

This test tries all the various forms of DDLs that create/alter/drop enums and makes sure they can be replicated automatically.

Success here is no errors and we end up with the same schemas (determined by YSQL dump ignoring OIDs) on both sides as well as the same enum label OID's.

Currently I have implemented "pass-through" where these DDLs are automatically executed on the target universe but I have not put in the code needed to keep the label OIDs the same.

I have thus disabled the test for now.

Note that the testing for the same enum label OIDs on both sides is done for all of these regression tests in case we end up adding a enum statement to one of those.

The BEGIN/ROLLBACK statements are commented out currently because there is a bug where we do not properly handle ordering DDLs inside of transactions; this should be fixed by #23957, as per the inserted TODO.
Jira: DB-12970

Test Plan:
I temporary added a enum statement to one of the other tests and the test successfully failed because the enum labels were different.

If I undisable the new test and comment out the OID perturbation, all the DDLs are successfully passed across and work.  If I leave in the perturbation, the test fails with backfill index crashing due to the OIDs being inconsistent between the pg_enum  table and the actual table's row contents.  This is expected -- we don't expect things to work unless the OIDs are the same.

Reviewers: jhe, xCluster, hsunder

Reviewed By: jhe

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D41267