[PLAT-15942]: "Force Delete" operation should be a from a separate More option
Summary:
Ignore errors and Force delete is bad behavior and it should be used only during exceptional cases. Why?This is because this flag ignores any errors during clean-up of the nodes and proceeds with deletion of the node with stale data on
that node.  We want to catch
these failures and have support look at those.

If we don't and allow such stale nodes to be available in the provider "free pool", we risk these nodes being used again for a different universe with stale data from the first universe thereby causing issues.Now, such risk is mitigated
now that we make the node go
in the decommissioned pool of the provider instead of the free pool of the provider on cleanup failures.

And then we re-run cleanup when trying to recommission the decommissioned nodes back to the free pool. But its still possible that there might be bugs in that code. So, we should only provide Ignore errors and Force delete as an advanced
option in the UI to be used
only when the user really cannot handle all the clean-up failures of the on-prem nodes.

Test Plan:
Please refer to screenshots
{F320006}

{F320007}

Closed the diff: D40116

Reviewers: jmak

Reviewed By: jmak

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40809