[PLAT-3739][xCluster] Add support for rotating TLS root certs while having xCluster replication
Summary:
This diff makes changes to update the source root certificate on the target universe when it is rotated on the source universe. Prior to this, rotating root certificated of the source universe would have caused a broken replication.

The automatic update happens only on universes where hot reload is available. For older universes, the user still has to restart replication after a root cert rotation.

Test Plan: Run a workload on the source universe and while the replication was going on, rotated the root certificated of the source universe. Ensured that the certificate is also updated on the target universe and observed the replication continues happening.

Reviewers: cwang, vbansal, nsingh, nbhatia

Reviewed By: cwang, nsingh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41143